+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 > This update runs a [significant migration](https://plugins.doublesecretagency.com/google-maps/address-field/multisite-support/#multisite-migration), which may notably increase the size of the `googlemaps_addresses` table for multisite projects. Please ensure that you have backed up the database before performing this migration.
 
 ### Added
-- [Multisite support](https://plugins.doublesecretagency.com/google-maps/address-field/multisite-support/). ([#11](https://github.com/doublesecretagency/craft-googlemaps/issues/11))
+- Added [multisite support](https://plugins.doublesecretagency.com/google-maps/address-field/multisite-support/). ([#11](https://github.com/doublesecretagency/craft-googlemaps/issues/11))
+- Added [`infoWindowData` option](https://plugins.doublesecretagency.com/google-maps/dynamic-maps/info-windows/#additional-variables-via-infowindowdata).
 
 ### Fixed
 - Fixed bug in “Test Address Lookup” utility.